requirement.txt
Package             Version
------------------- -------
asgiref             3.8.1
Django              5.0.7
djangorestframework 3.15.2
pip                 23.2.1
PyJWT               2.8.0
setuptools          68.2.2
sqlparse            0.5.0
typing_extensions   4.12.2
wheel               0.41.2
